Got more styling in place for the cart page and chooser stuff
- The root page now has a chooser and a cart, and some logic to determine where you go.
- Side-ported the Card component in and made some changes to make it work within the UE framework
- Copied over pxToRem too
- Got the bones set back up for the cart page

Doesn't yet pull cart data, but that's arguably the easy thing.

Some of the stuff in here should really go into smoot-design but my plan is to get them roughed in here first and debugged to an extent, then throw them in there (so I'm not waiting on reviews over there so much).